requirements.txt
Flask==3.1.0
pyserial==3.5
Flask-SocketIO==5.4.1